Obtain all issues of a JIRA query as a data.frame

Calls JIRA's latest REST API, optionally with basic authentication, to get all issues of a JIRA query (JQL). Allows to specify which fields to obtain.

JiraQuery2R(domain, user = NULL, password = NULL, query,
  fields = NULL, maxResults = NULL, verbose = FALSE)

Arguments

domain
Custom JIRA domain URL as for example https://bitvoodoo.atlassian.net.

user
Username used to authenticate access to JIRA your domain. If both username and password are not passed no authentication is made and only public domains can bet accesed.

password
Password used to authenticate access to JIRA your domain. If both username and password are not passed no authentication is made and only public domains can bet accesed.

query
JIRA's decoded JQL query. By definition, it works with with **Fields**, **Operators**, **Keywords** and **Functions**. To learn how to create a query visit this ATLASSIAN site.

fields
Optional argument to define the specific JIRA fields to obtain. If no value is entered, by defualt the following fields are passed: 'status','priority','created','reporter','summary','description','assignee','updated','issuetype','fixVersions'.

maxResults
Max results authorized to obtain for each API call. By default JIRA sets this value to 50 issues.

verbose
Gives the user the ability to have a verbose response from the JIRA API call.

Value

Returns a formatted data.frame with the issues according to the JQL query.

Warning

The function works with the latest JIRA REST API and to work you need to have a internet connection. Calling the function too many times might block your access and you will have to access manually online and enter a CAPTCHA at jira.enterprise.com/secure/Dashboard.jspa

See also

Examples

JiraQuery2R(domain = "https://bitvoodoo.atlassian.net", query = 'project="Congrats for Confluence"')

#>        id                                                        self
#>  1: 57446 https://bitvoodoo.atlassian.net/rest/api/latest/issue/57446
#>  2: 57383 https://bitvoodoo.atlassian.net/rest/api/latest/issue/57383
#>  3: 57327 https://bitvoodoo.atlassian.net/rest/api/latest/issue/57327
#>  4: 57249 https://bitvoodoo.atlassian.net/rest/api/latest/issue/57249
#>  5: 57157 https://bitvoodoo.atlassian.net/rest/api/latest/issue/57157
#>  6: 57041 https://bitvoodoo.atlassian.net/rest/api/latest/issue/57041
#>  7: 56904 https://bitvoodoo.atlassian.net/rest/api/latest/issue/56904
#>  8: 56797 https://bitvoodoo.atlassian.net/rest/api/latest/issue/56797
#>  9: 56796 https://bitvoodoo.atlassian.net/rest/api/latest/issue/56796
#> 10: 56795 https://bitvoodoo.atlassian.net/rest/api/latest/issue/56795
#> 11: 55800 https://bitvoodoo.atlassian.net/rest/api/latest/issue/55800
#> 12: 52800 https://bitvoodoo.atlassian.net/rest/api/latest/issue/52800
#> 13: 51806 https://bitvoodoo.atlassian.net/rest/api/latest/issue/51806
#> 14: 50050 https://bitvoodoo.atlassian.net/rest/api/latest/issue/50050
#>             key      status priority    created      reporter displayName
#>  1: CONGRATS-29 In Progress   Normal 2019-03-25         Stefanie Neidhart
#>  2: CONGRATS-27        Open   Normal 2019-02-04            Sascha Häusler
#>  3: CONGRATS-26      Closed   Normal 2018-11-07            Sascha Häusler
#>  4: CONGRATS-24 In Progress    Major 2018-09-12             Robin Stohler
#>  5: CONGRATS-23      Closed    Minor 2018-07-03 Niklas Becker [bitvoodoo]
#>  6: CONGRATS-20      Closed    Major 2018-03-19 Niklas Becker [bitvoodoo]
#>  7: CONGRATS-18      Closed    Minor 2017-12-05 Niklas Becker [bitvoodoo]
#>  8: CONGRATS-17        Open    Major 2017-09-26           Oliver Strässer
#>  9: CONGRATS-16        Open    Major 2017-09-26           Oliver Strässer
#> 10: CONGRATS-15        Open    Major 2017-09-26           Oliver Strässer
#> 11: CONGRATS-11      Closed    Major 2017-04-05          Daniele Talerico
#> 12:  CONGRATS-6      Closed    Major 2016-11-24              Sharon Funke
#> 13:  CONGRATS-3      Closed    Major 2016-10-21              Sharon Funke
#> 14:  CONGRATS-1      Closed    Major 2016-08-09           Oliver Strässer
#>                     reporter name    reporter displayName 1
#>  1: stefanie.fischer@bitvoodoo.ch         Stefanie Neidhart
#>  2:  sascha.haeusler@bitvoodoo.ch            Sascha Häusler
#>  3:  sascha.haeusler@bitvoodoo.ch            Sascha Häusler
#>  4:                         Robin             Robin Stohler
#>  5:                        Niklas Niklas Becker [bitvoodoo]
#>  6:                        Niklas Niklas Becker [bitvoodoo]
#>  7:                        Niklas Niklas Becker [bitvoodoo]
#>  8: oliver.straesser@bitvoodoo.ch           Oliver Strässer
#>  9: oliver.straesser@bitvoodoo.ch           Oliver Strässer
#> 10: oliver.straesser@bitvoodoo.ch           Oliver Strässer
#> 11: daniele.talerico@bitvoodoo.ch          Daniele Talerico
#> 12:     sharon.funke@bitvoodoo.ch              Sharon Funke
#> 13:     sharon.funke@bitvoodoo.ch              Sharon Funke
#> 14: oliver.straesser@bitvoodoo.ch           Oliver Strässer
#>                                                            summary
#>  1: Display issue of standard profile picture in Internet Explorer
#>  2:                            Congrats - Define performance tests
#>  3:                                 Congrats Data Center Checklist
#>  4:              Congrats for Confluence Data Center compatibility
#>  5:   If max entries is above 100 user icons overlap with Congrats
#>  6:      "You already congratulated" message missing after refresh
#>  7:      Add a dialogue for users that urges them to fill in dates
#>  8:                   Synchronisation with the //Seibert/Media CUP
#>  9:                 Add an Interface  to configure the sync fields
#> 10:                        Synchronisation with the Communardo UPP
#> 11:                       Display of age for birthday configurable
#> 12:                Do not display inactive users in Congrats Macro
#> 13:                         Incomplete rendering if placed in tabs
#> 14:                            Display current event in the center
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            description
#>  1:                                                                                                                                                                                               Steps to reproduce:\r\nEnv: Confluence 6.13.0, Internet Explorer 11\r\n* Person (who doesn't have an own profile picture set) has an occasion (birthday) today\r\n* Congrats Macro is added to a confluence page\r\n\r\nExpected result:\r\nStandard profile picture is shown correctly \r\n\r\nActual result:\r\nStandard profile picture is shown but has the wrong height:\r\n !image-2019-03-25-12-55-02-743.png|thumbnail! \r\n
#>  2:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               <NA>
#>  3:                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               <NA>
#>  4:                                 h2.Is Congrats for Confluence Data Center compatible?\r\nYes, Congrats for Confluence is Data Center ready but Atlassian removed the previous "Datacenter ready" Badge on 3 September 2018 and introduced the new program for Data Center approved apps. [Find more information here.|https://www.atlassian.com/licensing/data-center-approved-apps]\r\n\r\nAll vendors are required to reevaluate their apps and provide a vast amount of testing data to get the status back.\r\n\r\nIf you would like to speed up the process, to get this app Data Center approved again, vote this ticket up!
#>  5: If max entries is above 100, the dialog for writing your congratulation overlaps with the user icons.\r\n\r\nWhat I do\r\n* I add the Congrats Macro to a Confluence Page\r\n* I set Max Entries to 110\r\n* I set Days before and/or Event history to at least 110\r\n* I save the page\r\n* Someone has birthday, I click congratulate\r\n* The browser window is not full size\r\n\r\nWhat I expect\r\n* The macro works\r\n* The Congrats screen appears and I can enter my congratulation\r\n\r\nWhat happens instead\r\n* The user Icons overlap with the dialog \r\n\r\n!Screen Shot 2018-07-03 at 10.59.03.png|thumbnail! 
#>  6:                        What I do:\r\n* I congratulate someone\r\n\r\nWhat I expect\r\n* The reminder I already congratulated is displayed\r\n* The reminder stays there even after refreshing / revisiting the page containing the macro\r\n\r\nWhat happens instead:\r\n* The reminder is not displayed after refreshing the page\r\n* It is possible to press the congratulate button again, after trying to send the message an error occurs\r\n\r\nFurther informations:\r\n* Chrome generates a console error (POST 400) during the error\r\n* This was tested in Confluence 6.2.4 and Confluence 6.7.2 in Chrome and Firefox
#>  7:                                                                                                                                                                                                                                                           It would be easier for Confluence Administrators if the users fill in their birthdates and other occasions by themselves.\r\n\r\nA dialog / screen for every user asking to add birthday inside the profile or even the dialogue, would be an useful extension to the Add-on.\r\n\r\nThe dialog should be optional, and upon activation it should appear for every user.
#>  8:                                                                                                                                                                                                                                                                                                The CUP Add-on is stored their synced values as an property of the user too.\r\nhttps://marketplace.atlassian.com/plugins/net.seibertmedia.plugin.confluence.cup/server/overview\r\n\r\nThey can be retrieved by the UserDetailsManager, only the correct key is needed (needs to be configured in the admin dialog in CONGRATS-16)
#>  9:                                                                                                                                                                                                                                                                                                                                              * How this should be is explained here: https://wiki.bitvoodoo.ch/x/PQe5Aw\r\n* This Issue is only for the implementation  of the configuration\r\n* The Names  of the user profile fields can be entered mannually, without using the 3rd party REST API -> this will be done later 
#> 10:                                                                                                                                                                                                   Implement a synchronisation for the Communardo User Profile add-on (https://marketplace.atlassian.com/plugins/de.communardo.confluence.plugins.userprofile/server/overview)\r\n\r\nIf you have the field identifier you can retrieve the UPP value like this:\r\nhttps://git.bitvoodoo.ch/projects/BVLIB/repos/confluence-commons-library/browse/src/main/java/ch/bitvoodoo/confluence/commons/search/SearchQueryParser.java#154
#> 11:                                                                                                                                                                                                                                                                                                                                                                                                                              Please implement a new configuration parameter for the macro:\r\n- display age: true or false\r\n\r\nThis features has impact on the congratulation mask of the macro, see  !snapshot.png|thumbnail! 
#> 12:                                                                                                                                                                                                                                                                                                               Inactive users will be displayed within the Congrats Macro if one of their occasions will come up. Because they are inactive it makes not really sense to display them there and suggest that other users can congratulate. What about integrating a check about whether a user is active when fetching the entries?
#> 13:               * Congrats Makro in einem Tab Element (vom ET) platziert\r\n* beim Laden der Seite anderes Tab Element aktiv\r\n* beim Wechsel zum Tab mit dem Congrats Makro scheinen die Ressourcen, vermutlich vor allem das JS nicht richtig geladen zu werden\r\n\r\nScreenshots:\r\n[^tabs_congrats_wrong.png] -> Congrats wird nicht richtig geladen\r\n[^tabs_congrats_right.png] - > Congrats wird richtig geladen, dieser Tab war bereits beim Laden der gesamten Seite aktiv. Wenn jetzt zwischen Tabs hin- und hergewechselt wird, rendert das Congrats korrekt\r\n\r\nStorage Format:\r\n[^tabs_macros.xml]\r\n\r\n\r\n
#> 14:                                                                                                                                                                                                                                                                                                                                                At the moment the first event is centered, independent of the event date. \r\nFor example: If the macro shows an history of old events (last 3 days), the oldest event is selected, not the event from today. \r\n\r\nImprovement: Display the next event centered, not the oldest.
#>     assignee displayName                 assignee name assignee displayName 1
#>  1:         Yuriy Bilash     yuriy.bilash@bitvoodoo.ch           Yuriy Bilash
#>  2:                 <NA>                          <NA>                   <NA>
#>  3:         Yuriy Bilash     yuriy.bilash@bitvoodoo.ch           Yuriy Bilash
#>  4:                 <NA>                          <NA>                   <NA>
#>  5:         Yuriy Bilash     yuriy.bilash@bitvoodoo.ch           Yuriy Bilash
#>  6:       Sascha Häusler  sascha.haeusler@bitvoodoo.ch         Sascha Häusler
#>  7:                 <NA>                          <NA>                   <NA>
#>  8:                 <NA>                          <NA>                   <NA>
#>  9:         Yuriy Bilash     yuriy.bilash@bitvoodoo.ch           Yuriy Bilash
#> 10:         Yuriy Bilash     yuriy.bilash@bitvoodoo.ch           Yuriy Bilash
#> 11:    Stefanie Neidhart stefanie.fischer@bitvoodoo.ch      Stefanie Neidhart
#> 12:                 <NA>                          <NA>                   <NA>
#> 13:       Sascha Häusler  sascha.haeusler@bitvoodoo.ch         Sascha Häusler
#> 14:      Oliver Strässer oliver.straesser@bitvoodoo.ch        Oliver Strässer
#>        updated   issuetype
#>  1: 2019-04-04         Bug
#>  2: 2019-02-04        Task
#>  3: 2019-01-15        Task
#>  4: 2019-03-06        Epic
#>  5: 2019-01-29         Bug
#>  6: 2019-01-29         Bug
#>  7: 2017-12-05 New Feature
#>  8: 2018-09-28       Story
#>  9: 2018-09-28        Task
#> 10: 2018-09-28 New Feature
#> 11: 2017-04-19        Task
#> 12: 2016-12-14 Improvement
#> 13: 2019-01-29       Story
#> 14: 2016-11-30 Improvement
#>                                                                                                           fixVersions
#>  1:                                                                                                              <NA>
#>  2:                                                                                                              <NA>
#>  3:                                                                                                              <NA>
#>  4:                                                                                                              <NA>
#>  5: 1.3.4, 2019-01-29, 26902, FALSE, https://bitvoodoo.atlassian.net/rest/api/2/version/26902, Several bugfixes, TRUE
#>  6: 1.3.4, 2019-01-29, 26902, FALSE, https://bitvoodoo.atlassian.net/rest/api/2/version/26902, Several bugfixes, TRUE
#>  7:                                                                                                              <NA>
#>  8:                                                                                                              <NA>
#>  9:                                                                                                              <NA>
#> 10:                                                                                                              <NA>
#> 11:                 , 1.3.0, 2017-04-18, 26300, FALSE, https://bitvoodoo.atlassian.net/rest/api/2/version/26300, TRUE
#> 12:                   1.2.2, 2016-12-14, 25600, FALSE, https://bitvoodoo.atlassian.net/rest/api/2/version/25600, TRUE
#> 13: 1.3.4, 2019-01-29, 26902, FALSE, https://bitvoodoo.atlassian.net/rest/api/2/version/26902, Several bugfixes, TRUE
#> 14:                             , 1.1.6, 24201, FALSE, https://bitvoodoo.atlassian.net/rest/api/2/version/24201, TRUE
#>     assignee
#>  1:       NA
#>  2:       NA
#>  3:       NA
#>  4:       NA
#>  5:       NA
#>  6:       NA
#>  7:       NA
#>  8:       NA
#>  9:       NA
#> 10:       NA
#> 11:       NA
#> 12:       NA
#> 13:       NA
#> 14:       NA