Obtain all projects as a data.frame

Calls JIRA's latest REST API to obtain all the basic project information (Name, Key, Id, Description, etc.)

Projects2R(domain, user = NULL, password = NULL, expand = NULL,
  verbose = FALSE)

Arguments

domain
Custom JIRA domain URL as for example https://bitvoodoo.atlassian.net

user
Username used to authenticate access to JIRA your domain. If both username and password are not passed no authentication is made and only public domains can bet accesed. Optional parameter.

password
Password used to authenticate access to JIRA your domain. If both username and password are not passed no authentication is made and only public domains can bet accesed. Optional parameter.

expand
Specific JIRA fields the user wants to obtain for a specific field. Optional parameter. By default

verbose
Gives the user the ability to have a verbose response from the JIRA API call.

Value

Returns a data.frame with a list of projects for which the user has the BROWSE, ADMINISTER or PROJECT_ADMIN project permission.

Warning

The function works with the JIRA REST API. Thus, to work it needs a internet connection. Calling the function too many times might block your access and you will have to access manually online and enter a CAPTCHA at jira.yourdomain.com/secure/Dashboard.jspa

See also

Examples

Projects2R("https://bitvoodoo.atlassian.net")

#>                                          expand
#>  1: description,lead,issueTypes,url,projectKeys
#>  2: description,lead,issueTypes,url,projectKeys
#>  3: description,lead,issueTypes,url,projectKeys
#>  4: description,lead,issueTypes,url,projectKeys
#>  5: description,lead,issueTypes,url,projectKeys
#>  6: description,lead,issueTypes,url,projectKeys
#>  7: description,lead,issueTypes,url,projectKeys
#>  8: description,lead,issueTypes,url,projectKeys
#>  9: description,lead,issueTypes,url,projectKeys
#> 10: description,lead,issueTypes,url,projectKeys
#> 11: description,lead,issueTypes,url,projectKeys
#> 12: description,lead,issueTypes,url,projectKeys
#> 13: description,lead,issueTypes,url,projectKeys
#> 14: description,lead,issueTypes,url,projectKeys
#> 15: description,lead,issueTypes,url,projectKeys
#> 16: description,lead,issueTypes,url,projectKeys
#> 17: description,lead,issueTypes,url,projectKeys
#> 18: description,lead,issueTypes,url,projectKeys
#> 19: description,lead,issueTypes,url,projectKeys
#>                                                         self    id         key
#>  1: https://bitvoodoo.atlassian.net/rest/api/2/project/12320 12320    MACRODOC
#>  2: https://bitvoodoo.atlassian.net/rest/api/2/project/12420 12420       THEME
#>  3: https://bitvoodoo.atlassian.net/rest/api/2/project/12520 12520         SBB
#>  4: https://bitvoodoo.atlassian.net/rest/api/2/project/12521 12521          SD
#>  5: https://bitvoodoo.atlassian.net/rest/api/2/project/12620 12620      CFSYNC
#>  6: https://bitvoodoo.atlassian.net/rest/api/2/project/13100 13100    TEMPBLOG
#>  7: https://bitvoodoo.atlassian.net/rest/api/2/project/12720 12720    REDIRECT
#>  8: https://bitvoodoo.atlassian.net/rest/api/2/project/13200 13200       LABEL
#>  9: https://bitvoodoo.atlassian.net/rest/api/2/project/11380 11380    PANELBOX
#> 10: https://bitvoodoo.atlassian.net/rest/api/2/project/12820 12820         REG
#> 11: https://bitvoodoo.atlassian.net/rest/api/2/project/13300 13300    LABELFIX
#> 12: https://bitvoodoo.atlassian.net/rest/api/2/project/13301 13301     NTCLOUD
#> 13: https://bitvoodoo.atlassian.net/rest/api/2/project/11382 11382    NAVITABS
#> 14: https://bitvoodoo.atlassian.net/rest/api/2/project/11383 11383    LANGUAGE
#> 15: https://bitvoodoo.atlassian.net/rest/api/2/project/12120 12120   SCHEDULER
#> 16: https://bitvoodoo.atlassian.net/rest/api/2/project/12920 12920    SYNCTEST
#> 17: https://bitvoodoo.atlassian.net/rest/api/2/project/11384 11384 VIEWTRACKER
#> 18: https://bitvoodoo.atlassian.net/rest/api/2/project/13020 13020    CONGRATS
#> 19: https://bitvoodoo.atlassian.net/rest/api/2/project/12220 12220     EXTLINK
#>                                        name
#>  1:      Macro Documentation for Confluence
#>  2:         Enterprise Theme for Confluence
#>  3:              SBB Widgets for Confluence
#>  4:                        Sequence Diagram
#>  5:        Custom Field Option Synchroniser
#>  6: Templates for Blog Posts for Confluence
#>  7:        Homepage Redirect for Confluence
#>  8:          Label Scheduler for Confluence
#>  9:      Advanced Panelboxes for Confluence
#> 10:   bitvoodoo Registration for Confluence
#> 11:                             Label Fixer
#> 12:    Navitabs - Tabs for Confluence Cloud
#> 13:          Navitabs - Tabs for Confluence
#> 14:          Language Macros for Confluence
#> 15:      Content Scheduler for Confluence\t
#> 16:                                SyncTest
#> 17:  Viewtracker - Analytics for Confluence
#> 18:                 Congrats for Confluence
#> 19:           External Links for Confluence
#>                                                                  avatarUrls 48x48
#>  1: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12320&avatarId=14609
#>  2: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12420&avatarId=14603
#>  3: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12520&avatarId=14610
#>  4: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12521&avatarId=11961
#>  5: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12620&avatarId=14606
#>  6: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=13100&avatarId=14611
#>  7: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12720&avatarId=14607
#>  8: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=13200&avatarId=14600
#>  9: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=11380&avatarId=14604
#> 10: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12820&avatarId=14702
#> 11: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=13300&avatarId=14700
#> 12: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=13301&avatarId=14703
#> 13: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=11382&avatarId=14500
#> 14: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=11383&avatarId=14608
#> 15: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12120&avatarId=14605
#> 16:           https://bitvoodoo.atlassian.net/secure/projectavatar?avatarId=13163
#> 17: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=11384&avatarId=14612
#> 18: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=13020&avatarId=14601
#> 19: https://bitvoodoo.atlassian.net/secure/projectavatar?pid=12220&avatarId=14602
#>                                                                             avatarUrls 24x24
#>  1: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12320&avatarId=14609
#>  2: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12420&avatarId=14603
#>  3: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12520&avatarId=14610
#>  4: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12521&avatarId=11961
#>  5: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12620&avatarId=14606
#>  6: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=13100&avatarId=14611
#>  7: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12720&avatarId=14607
#>  8: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=13200&avatarId=14600
#>  9: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=11380&avatarId=14604
#> 10: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12820&avatarId=14702
#> 11: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=13300&avatarId=14700
#> 12: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=13301&avatarId=14703
#> 13: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=11382&avatarId=14500
#> 14: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=11383&avatarId=14608
#> 15: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12120&avatarId=14605
#> 16:           https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&avatarId=13163
#> 17: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=11384&avatarId=14612
#> 18: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=13020&avatarId=14601
#> 19: https://bitvoodoo.atlassian.net/secure/projectavatar?size=small&pid=12220&avatarId=14602
#>                                                                              avatarUrls 16x16
#>  1: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12320&avatarId=14609
#>  2: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12420&avatarId=14603
#>  3: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12520&avatarId=14610
#>  4: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12521&avatarId=11961
#>  5: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12620&avatarId=14606
#>  6: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=13100&avatarId=14611
#>  7: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12720&avatarId=14607
#>  8: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=13200&avatarId=14600
#>  9: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=11380&avatarId=14604
#> 10: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12820&avatarId=14702
#> 11: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=13300&avatarId=14700
#> 12: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=13301&avatarId=14703
#> 13: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=11382&avatarId=14500
#> 14: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=11383&avatarId=14608
#> 15: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12120&avatarId=14605
#> 16:           https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&avatarId=13163
#> 17: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=11384&avatarId=14612
#> 18: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=13020&avatarId=14601
#> 19: https://bitvoodoo.atlassian.net/secure/projectavatar?size=xsmall&pid=12220&avatarId=14602
#>                                                                              avatarUrls 32x32
#>  1: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12320&avatarId=14609
#>  2: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12420&avatarId=14603
#>  3: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12520&avatarId=14610
#>  4: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12521&avatarId=11961
#>  5: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12620&avatarId=14606
#>  6: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=13100&avatarId=14611
#>  7: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12720&avatarId=14607
#>  8: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=13200&avatarId=14600
#>  9: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=11380&avatarId=14604
#> 10: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12820&avatarId=14702
#> 11: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=13300&avatarId=14700
#> 12: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=13301&avatarId=14703
#> 13: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=11382&avatarId=14500
#> 14: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=11383&avatarId=14608
#> 15: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12120&avatarId=14605
#> 16:           https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&avatarId=13163
#> 17: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=11384&avatarId=14612
#> 18: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=13020&avatarId=14601
#> 19: https://bitvoodoo.atlassian.net/secure/projectavatar?size=medium&pid=12220&avatarId=14602
#>     projectTypeKey simplified   style isPrivate
#>  1:       software      FALSE classic     FALSE
#>  2:       software      FALSE classic     FALSE
#>  3:       software      FALSE classic     FALSE
#>  4:       software      FALSE classic     FALSE
#>  5:       software      FALSE classic     FALSE
#>  6:       software      FALSE classic     FALSE
#>  7:       software      FALSE classic     FALSE
#>  8:       software      FALSE classic     FALSE
#>  9:       software      FALSE classic     FALSE
#> 10:       software      FALSE classic     FALSE
#> 11:       software      FALSE classic     FALSE
#> 12:       software      FALSE classic     FALSE
#> 13:       software      FALSE classic     FALSE
#> 14:       software      FALSE classic     FALSE
#> 15:       software      FALSE classic     FALSE
#> 16:       software      FALSE classic     FALSE
#> 17:       software      FALSE classic     FALSE
#> 18:       software      FALSE classic     FALSE
#> 19:       software      FALSE classic     FALSE