Retrieves all dashboards a data.frame

Calls JIRA's latest REST API, optionally with basic authentication, to get all dashboards

get_jira_dashboards(
  domain = NULL,
  username = NULL,
  password = NULL,
  maxResults = 20L,
  verbose = FALSE
)

Arguments

domain
Custom JIRA domain URL as for example https://bugreports.qt.io. Can be passed as a parameter or can be previously defined through the save_jira_credentials() function.

username
Username used to authenticate the access to the JIRA domain. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.

password
Password used to authenticate the access to the JIRA domain. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.

maxResults
Max results authorized to obtain for each API call. By default JIRA sets this value to 20 issues.

verbose
Explicitly informs the user of the JIRA API request process.

Value

Returns a flattened, formatted data.frame with the dashboards in the domain.

See also

Examples

get_jira_dashboards(domain = "https://bugreports.qt.io")

#>       id
#> 1  15113
#> 2  15116
#> 3  15618
#> 4  15620
#> 5  16710
#> 6  17915
#> 7  15112
#> 8  17916
#> 9  17917
#> 10 17918
#> 11 17919
#> 12 12514
#> 13 14918
#> 14 13218
#> 15 16212
#> 16 11521
#> 17 10898
#> 18 11125
#> 19 15413
#> 20 10517
#> 21 14310
#> 22 11315
#> 23 15918
#> 24 17022
#> 25 11005
#> 26 13216
#> 27 11191
#> 28 11511
#> 29 13916
#> 30 12911
#> 31 10364
#> 32 14529
#> 33 11513
#> 34 11817
#> 35 14517
#> 36 13220
#> 37 13214
#> 38 11118
#> 39 10778
#> 40 11130
#> 41 12116
#> 42 11820
#> 43 11093
#> 44 12512
#> 45 12010
#> 46 14524
#> 47 17811
#> 48 17021
#> 49 13912
#> 50 11080
#> 51 13712
#> 52 15911
#> 53 13310
#> 54 10000
#> 55 14510
#> 56 15714
#> 57 11712
#> 58 10161
#> 59 15315
#>                                                                                                 name
#> 1                                                                                   5.10 Infra Items
#> 2                                                                                   5.11 Infra Items
#> 3                                                                                   5.12 Infra items
#> 4                                                                                   5.13 Infra items
#> 5                                                                                   5.14 Infra items
#> 6                                                                                   5.15 Infra items
#> 7                                                                                    5.9 Infra Items
#> 8                                                                                    6.0 Infra items
#> 9                                                                                    6.1 Infra items
#> 10                                                                                   6.2 Infra items
#> 11                                                                                   6.3 Infra items
#> 12                                                                           Accessibility Dashboard
#> 13                                                                                         Autotests
#> 14                                                                   Bug Fix Week: Subgroups of Bugs
#> 15                                                                            Build System Dashboard
#> 16                                                                              Documentation issues
#> 17                                                                                        Earth Team
#> 18                                                                             Earth Team Statistics
#> 19                                                                          Embedded Linux / Wayland
#> 20                                                                            Embedded Team Overview
#> 21                                                                                               f5k
#> 22                                                                        GEG - Gerrit triaging view
#> 23                                                                                            Gerrit
#> 24                                                                                          High-DPI
#> 25                                                                                               hjk
#> 26                                                                                           Jessica
#> 27 Meego Harmattan Api - qmake.exe - query is nor parseable error while installing qt sdk on windows
#> 28                                                                                       My overview
#> 29                                                                                            PTHDPI
#> 30                                                                                            PySide
#> 31                                                                                        Qt - Focus
#> 32                                                                          Qt 3D Overview Dashboard
#> 33                                                                                      Qt 5 - Focus
#> 34                                                                              Qt 5.x bug dashboard
#> 35                                                                                      Qt autotests
#> 36                                                                     Qt Bug Fix Week: Overall view
#> 37                                                                      Qt bug fixing week dashboard
#> 38                                                                                        Qt Creator
#> 39                                                                              Qt Creator Dashboard
#> 40                                                                                  Qt for Harmattan
#> 41                                                                                      Qt for Tizen
#> 42                                                                  Qt Installer Framework Dashboard
#> 43                                                                           Qt Multimedia Dashboard
#> 44                                                                              Qt Network Dashboard
#> 45                                                                             Qt QA Infra Dashboard
#> 46                                                                               Qt Quick Controls 2
#> 47                                                              Qt Safe Renderer 1.2 Jira statistics
#> 48                                                                            Qt Team Platforms Area
#> 49                                                                            Qt WebEngine Dashboard
#> 50                                                                                Qt Webkit OverView
#> 51                                                                                     Quick and QML
#> 52                                                                                         Releasing
#> 53                                                                                       Rottin Dash
#> 54                                                                                  System Dashboard
#> 55                                                                                            Triage
#> 56                                                                                                TV
#> 57                                                                    Unresolved Qt 5.0 bug overview
#> 58                                                                                           Usefull
#> 59                                                                                       WebAssembly
#>                                                   self
#> 1  https://bugreports.qt.io/rest/api/2/dashboard/15113
#> 2  https://bugreports.qt.io/rest/api/2/dashboard/15116
#> 3  https://bugreports.qt.io/rest/api/2/dashboard/15618
#> 4  https://bugreports.qt.io/rest/api/2/dashboard/15620
#> 5  https://bugreports.qt.io/rest/api/2/dashboard/16710
#> 6  https://bugreports.qt.io/rest/api/2/dashboard/17915
#> 7  https://bugreports.qt.io/rest/api/2/dashboard/15112
#> 8  https://bugreports.qt.io/rest/api/2/dashboard/17916
#> 9  https://bugreports.qt.io/rest/api/2/dashboard/17917
#> 10 https://bugreports.qt.io/rest/api/2/dashboard/17918
#> 11 https://bugreports.qt.io/rest/api/2/dashboard/17919
#> 12 https://bugreports.qt.io/rest/api/2/dashboard/12514
#> 13 https://bugreports.qt.io/rest/api/2/dashboard/14918
#> 14 https://bugreports.qt.io/rest/api/2/dashboard/13218
#> 15 https://bugreports.qt.io/rest/api/2/dashboard/16212
#> 16 https://bugreports.qt.io/rest/api/2/dashboard/11521
#> 17 https://bugreports.qt.io/rest/api/2/dashboard/10898
#> 18 https://bugreports.qt.io/rest/api/2/dashboard/11125
#> 19 https://bugreports.qt.io/rest/api/2/dashboard/15413
#> 20 https://bugreports.qt.io/rest/api/2/dashboard/10517
#> 21 https://bugreports.qt.io/rest/api/2/dashboard/14310
#> 22 https://bugreports.qt.io/rest/api/2/dashboard/11315
#> 23 https://bugreports.qt.io/rest/api/2/dashboard/15918
#> 24 https://bugreports.qt.io/rest/api/2/dashboard/17022
#> 25 https://bugreports.qt.io/rest/api/2/dashboard/11005
#> 26 https://bugreports.qt.io/rest/api/2/dashboard/13216
#> 27 https://bugreports.qt.io/rest/api/2/dashboard/11191
#> 28 https://bugreports.qt.io/rest/api/2/dashboard/11511
#> 29 https://bugreports.qt.io/rest/api/2/dashboard/13916
#> 30 https://bugreports.qt.io/rest/api/2/dashboard/12911
#> 31 https://bugreports.qt.io/rest/api/2/dashboard/10364
#> 32 https://bugreports.qt.io/rest/api/2/dashboard/14529
#> 33 https://bugreports.qt.io/rest/api/2/dashboard/11513
#> 34 https://bugreports.qt.io/rest/api/2/dashboard/11817
#> 35 https://bugreports.qt.io/rest/api/2/dashboard/14517
#> 36 https://bugreports.qt.io/rest/api/2/dashboard/13220
#> 37 https://bugreports.qt.io/rest/api/2/dashboard/13214
#> 38 https://bugreports.qt.io/rest/api/2/dashboard/11118
#> 39 https://bugreports.qt.io/rest/api/2/dashboard/10778
#> 40 https://bugreports.qt.io/rest/api/2/dashboard/11130
#> 41 https://bugreports.qt.io/rest/api/2/dashboard/12116
#> 42 https://bugreports.qt.io/rest/api/2/dashboard/11820
#> 43 https://bugreports.qt.io/rest/api/2/dashboard/11093
#> 44 https://bugreports.qt.io/rest/api/2/dashboard/12512
#> 45 https://bugreports.qt.io/rest/api/2/dashboard/12010
#> 46 https://bugreports.qt.io/rest/api/2/dashboard/14524
#> 47 https://bugreports.qt.io/rest/api/2/dashboard/17811
#> 48 https://bugreports.qt.io/rest/api/2/dashboard/17021
#> 49 https://bugreports.qt.io/rest/api/2/dashboard/13912
#> 50 https://bugreports.qt.io/rest/api/2/dashboard/11080
#> 51 https://bugreports.qt.io/rest/api/2/dashboard/13712
#> 52 https://bugreports.qt.io/rest/api/2/dashboard/15911
#> 53 https://bugreports.qt.io/rest/api/2/dashboard/13310
#> 54 https://bugreports.qt.io/rest/api/2/dashboard/10000
#> 55 https://bugreports.qt.io/rest/api/2/dashboard/14510
#> 56 https://bugreports.qt.io/rest/api/2/dashboard/15714
#> 57 https://bugreports.qt.io/rest/api/2/dashboard/11712
#> 58 https://bugreports.qt.io/rest/api/2/dashboard/10161
#> 59 https://bugreports.qt.io/rest/api/2/dashboard/15315
#>                                                                 view
#> 1  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15113
#> 2  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15116
#> 3  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15618
#> 4  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15620
#> 5  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=16710
#> 6  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17915
#> 7  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15112
#> 8  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17916
#> 9  https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17917
#> 10 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17918
#> 11 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17919
#> 12 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=12514
#> 13 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=14918
#> 14 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13218
#> 15 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=16212
#> 16 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11521
#> 17 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=10898
#> 18 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11125
#> 19 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15413
#> 20 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=10517
#> 21 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=14310
#> 22 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11315
#> 23 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15918
#> 24 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17022
#> 25 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11005
#> 26 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13216
#> 27 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11191
#> 28 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11511
#> 29 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13916
#> 30 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=12911
#> 31 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=10364
#> 32 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=14529
#> 33 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11513
#> 34 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11817
#> 35 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=14517
#> 36 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13220
#> 37 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13214
#> 38 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11118
#> 39 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=10778
#> 40 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11130
#> 41 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=12116
#> 42 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11820
#> 43 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11093
#> 44 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=12512
#> 45 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=12010
#> 46 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=14524
#> 47 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17811
#> 48 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=17021
#> 49 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13912
#> 50 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11080
#> 51 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13712
#> 52 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15911
#> 53 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=13310
#> 54 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=10000
#> 55 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=14510
#> 56 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15714
#> 57 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=11712
#> 58 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=10161
#> 59 https://bugreports.qt.io/secure/Dashboard.jspa?selectPageId=15315