Get JIRA server groups data.frame

Makes a request to JIRA's latest REST API to retrieve all the groups in JIRA.

get_jira_groups(
  domain = NULL,
  username = NULL,
  password = NULL,
  verbose = FALSE,
  maxResults = 1000
)

Arguments

domain
Custom JIRA domain URL as for example https://bitvoodoo.atlassian.net. Can be passed as a parameter or can be previously defined through the save_jira_credentials() function.

username
Username used to authenticate the access to the JIRA domain. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.

password
Password used to authenticate the access to the JIRA domain. If both username and password are not passed no authentication is made and only public domains can bet accessed. Optional parameter.

verbose
Explicitly informs the user of the JIRA API request process.

maxResults
Number of maximum groups to return. Set by default to 1000.

Value

Returns a data.frame with all the JIRA groups

Warning

The function works with the JIRA REST API. Thus, to work it needs an internet connection. Calling the function too many times might block your access and you will have to access manually online and enter a CAPTCHA at jira.yourdomain.com/secure/Dashboard.jspa.

See also

Examples

get_jira_groups("https://bitvoodoo.atlassian.net")

#>                                                  name
#> 1                                         access-jira
#> 2                                      administrators
#> 3                              atlassian-addons-admin
#> 4                                   confluence-admins
#> 5                                    confluence-users
#> 6                                          developers
#> 7                                 jira-administrators
#> 8                                      jira-ops-users
#> 9                                          jira-users
#> 10                                     opsgenie-users
#> 11                                        site-admins
#> 12                                 site-admins-backup
#> 13                              system-administrators
#> 14 trusted-users-b6b9b051-e45d-494a-832d-0b5a8bb56619
#> 15                                              users
#>                                                  html labels
#> 1                                         access-jira   <NA>
#> 2                                      administrators   <NA>
#> 3                              atlassian-addons-admin   <NA>
#> 4                                   confluence-admins   <NA>
#> 5                                    confluence-users   <NA>
#> 6                                          developers   <NA>
#> 7                                 jira-administrators   <NA>
#> 8                                      jira-ops-users   <NA>
#> 9                                          jira-users   <NA>
#> 10                                     opsgenie-users   <NA>
#> 11                                        site-admins   <NA>
#> 12                                 site-admins-backup   <NA>
#> 13                              system-administrators   <NA>
#> 14 trusted-users-b6b9b051-e45d-494a-832d-0b5a8bb56619   <NA>
#> 15                                              users   <NA>
#>                                 groupId
#> 1  bb5faf71-0c2c-4d05-a0b6-b1db5c1f39ed
#> 2  0df36cb7-61e8-40cc-bfed-262ca2362a74
#> 3  1e5d93c3-8760-4124-8f4d-fa669c4a3263
#> 4  aa1353cf-aa68-4646-b7ca-0a87e467a2cc
#> 5  8969657d-d56f-40a1-9000-7f6a8856dccd
#> 6  67fcbdda-fea6-405b-a2be-dd771065ec88
#> 7  d0b56c93-84b6-4db3-9e5d-b55b49cc789d
#> 8  944ca2a2-cad6-4ff3-8c21-e94731c5b655
#> 9  3799fc5c-8e4d-4caf-8e2a-9e502cf3371c
#> 10 df6e3815-18c0-4af2-bcef-7e1889f4bee3
#> 11 2db77824-5901-43ff-aabd-0b8e493a720f
#> 12 6fc0e67a-1798-466f-a189-d387a7973706
#> 13 eff41464-a7d2-4378-ad65-8224231b7b09
#> 14 efa386a9-4523-4515-9555-27ed1120d7e9
#> 15 1592ff4a-fad5-487c-a412-c1d71019a3d0